Fix building n-quads
This is essentially editing the regexp to make it accept quoted double quotes,
and correcting the quads line function to use the graph regexp specified.

It may not be a proper solution because I couldn't work out how to add more
characters to the OR statement without more nested brackets (this actually
adds one more nested bracket). This works for now but the library presumably
isn't standard.